Properly detect final file extension.
A file named `foo.bar.baz` should use a loader for `baz`, not `bar.baz`.

Closes #31
