+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ unreleased
 
 * Removed zope.interface requirement, since the interfaces themselves were removed in 0.2.0.
 * Removed useless console_script entry point.
+* Changed loader detection to only consider the final extension in a filename.
 
 0.2.0 (2015-06-14)
 -----------------------------------------
